did orderService
finished orderServices

+++ src/main/java/com/michelle/Kali/models/User.java
@@ -28,7 +28,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class User {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     
     @Email
     @Pattern(regexp=".+@.+\\..+")
@@ -64,12 +64,14 @@ public class User {
     	this.email = email;
     	this.password = password;
     }
+    
+    
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -153,8 +155,5 @@ public class User {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-    
-    
-    
-    
+
 }
